Move QuerySettings out of Rubber into Context
